Added operationId parameter to OpenAPI Metadata Object and implemented test case to ensure proper usage in generated openApi description file.
Reason for change:
Many code generators primarily use the operationId for generating client side method names. Setting this field allows to control how the generated code on the client side looks like. Also it becomes possible to change the default method names. This is quite handy if the default names are not meaningful enough or must be overridden for legacy reasons.

Limitations:
According to the documentation, the operationId must be unique within a api. The code does not contain any checks to ensure uniqueness. The developer must carefully assign operationIds to avoid naming conflicts